solve the trivial.eo exponential run time and break mutability:
The trivial.eo implements a loop, which iterates ~10 times and executes several steps in each iteration. This is done via continuation: a step is implemented as a separate function, so function tmpFun0 calls tmpFun1, which calls tmpFun2, ... tmpFun13 calls tmpFun3 and the cycle continues until some of them returns.

The problem is that run time of a function rises exponentially with time, so, say tmpFun5 works for ~5min, tmpFun10 for ~10hrs and tmpFun13 is never reached.

Probably, this happens because some computations are performed more that once. This problem is already partialy solved: the generated .java wraps each AtComposite(anonymousFunction) construct into AtOnce, which memoizes a result of the AtComposite. This memoization is not implemented in the runtime code though, so some pattern of usage of array.mapi/array.get/int.less (like in trivial.eo) leads to exponential run time, where the run time should be linear.

This commit adds memoization to the runtime library. It solves the problem, but breaks (of course) all the mutable code, that is, usages of the memory and random objects.

Additional context: trivial.eo is a result of translation of https://github.com/polystat/py2eo/blob/6783fd11c947dc1fc1836e1e3edcca3d34020b4f/python/python3/test/trivial.py
That is, it computes a Fibonacci number.